+++ packages/babel-plugin/src/__test__/__fixtures__/base/output.js
@@ -1,3 +1,4 @@
+import { Box as __Box } from "@kuma-ui/core";
 import React from "react";
 import { css } from "@kuma-ui/core";
 const style = "kuma-3206633536";
